python_scripts/requirements.txt
pymongo
pymupdf
sentence-transformers
faiss-cpu
numpy
requests
scikit-learn
torch
fastapi
uvicorn
